complete multi and single select filter ui for regular view
