--- src/components/FilterDropdown/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./FilterDropdown";
-
