+++ src/pages/Timeline/TimelineFilter.css
@@ -1,3 +1,11 @@
+.timeline-filter {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    width: min(90%, 1000px);
+}
+
 .filter-dropdown {
     display: flex;
     align-items: center;
@@ -10,10 +18,6 @@
     margin: 0 5px;
 }
 
-.filter {
-    display: flex;
-}
-
 .filter-divider {
     content: '';
     display: inline-block;
@@ -23,8 +27,6 @@
     border: 0.7px solid #e2e2e2;
 }
 
-.filter-dropdown-text {
-    display: inline-block;
-    color: #7e7e7e;
-    margin-right: 10px;
+@media only screen and (max-width: 1123px) {
+	
 }
