--- src/pages/Timeline/TimelineFilterContainer.css
@@ -1,22 +1,0 @@
-.timeline-filter-container {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    width: min(90%, 1000px);
-    /* margin: 0 auto;
-    text-align: left; */
-}
-
-.add-update-button {
-    padding: 5px 10px;
-    margin: 2px;
-    /* margin-left: 65px; */
-    background-color: #4a668a;
-    border-radius: 10px;
-}
-
-
-@media only screen and (max-width: 1123px) {
-	
-}
